+++ package.json
@@ -4,6 +4,7 @@
   "private": true,
   "dependencies": {
     "firebase": "^4.8.1",
+    "firebase-admin": "^5.12.0",
     "react": "^16.2.0",
     "react-dom": "^16.2.0",
     "react-scripts": "1.0.17",
@@ -14,5 +15,8 @@
     "build": "react-scripts build",
     "test": "react-scripts test --env=jsdom",
     "eject": "react-scripts eject"
+  },
+  "devDependencies": {
+    "enzyme": "^3.3.0"
   }
 }
