+++ src/components/App.js
@@ -1,0 +1,8 @@
+import React from "react";
+import "../styles/App.css";
+
+const App = (props) => (
+  <p>hi</p>
+);
+
+export default App;
